% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{add_data}
\alias{add_data}
\title{Add training and testing data.}
\usage{
add_data(dt.ad, training_data, testing_data)
}
\arguments{
\item{dt.ad}{An S3 data structure used in the package, see \code{\link{Summary}}.}

\item{training_data}{A data.table with \code{dt.ad$param$num_variables} number of columns.}

\item{testing_data}{A data.table with \code{dt.ad$param$num_variables} number of columns.}
}
\value{
An S3 object of mvad type.
}
\description{
\code{add_data} returns the added anomaly detection data structure
with the training and testing data provided as arguments.
}
\examples{
dt.ad <- add_data(setup_model(),
                  data.table("v" = 1:10,"i" = 11:20, "t" = 21:30, "f" = 31:40),
                  data.table("v" = 1:5,"i" = 6:10, "t" = 11:15, "f" = 16:20))
}
