% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{detrend_data}
\alias{detrend_data}
\title{internal}
\usage{
detrend_data(
  data.preprocessed,
  data_to_use = data.preprocessed$raw_testing_data
)
}
\arguments{
\item{data.preprocessed}{MVAD obj}

\item{data_to_use}{data.table}
}
\description{
Not to be used for external
}
