% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{evaluate}
\alias{evaluate}
\title{Wrapper to evaluate all steps of anomaly detection.}
\usage{
evaluate(dt.ad)
}
\arguments{
\item{dt.ad}{An S3 object of MVAD type (partially evaluated)}
}
\value{
An S3 object of MVAD type (fully evaluated).
}
\description{
\code{evaluate} runs preprocessing, model fitting, residual finding, and anomaly scoring.
Before calling, \code{\link{setup_model}}, and\code{\link{add_data}} should be called.
}
