% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{get_anomaly_score}
\alias{get_anomaly_score}
\title{Scores data residuals using model.}
\usage{
get_anomaly_score(dt.ad)
}
\arguments{
\item{dt.ad}{- An S3 object of MVAD type (partially evaluated)}
}
\value{
An S3 object of MVAD type (fully evaluated).
}
\description{
\code{get_anomaly_detection} returns the scores of the residuals calculated by the models parameters.
Before calling, \code{\link{setup_model}}, \code{\link{add_data}}, \code{\link{preprocess_data}},
\code{\link{train_model}}, and \code{\link{get_residual}} should be called.
}
