% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{get_residual}
\alias{get_residual}
\title{Gets data residuals using model and testing data.}
\usage{
get_residual(dt.ad)
}
\arguments{
\item{dt.ad}{- An S3 object of MVAD type (partially evaluated)}
}
\value{
An S3 object of MVAD type (partially evaluated).
}
\description{
\code{get_residual} returns the residuals calculated by the models prediction and testing data.
Before calling, \code{\link{setup_model}}, \code{\link{add_data}}, \code{\link{preprocess_data}},
and \code{\link{train_model}} should be called.
}
