% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{get_test_data}
\alias{get_test_data}
\title{Create testing data synthetically.}
\usage{
get_test_data()
}
\value{
A data.table with 4 column/variables and 1.1 * tau rows.
}
\description{
\code{get_test_data} returns example multivartiate testing data with similar trend and standard deviations
as (a) real data set did
}
\examples{
dt.synthetic_data <- get_test_data()

}
