% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{linearize_data}
\alias{linearize_data}
\title{internal}
\usage{
linearize_data(dt.ad, coarse_data)
}
\arguments{
\item{dt.ad}{MVAD obj}

\item{coarse_data}{data.table}
}
\description{
Not to be used for external
}
