% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{make_coarse}
\alias{make_coarse}
\title{internal}
\usage{
make_coarse(dt.ad, training_data)
}
\arguments{
\item{dt.ad}{MVAD obj}

\item{training_data}{data.frame}
}
\description{
Not to be used for external
}
