% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{normalize_data}
\alias{normalize_data}
\title{internal}
\usage{
normalize_data(dt.ad, linearized_data)
}
\arguments{
\item{dt.ad}{MVAD obj}

\item{linearized_data}{data.table}
}
\description{
Not to be used for external
}
