% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess Data.}
\usage{
preprocess_data(dt.ad)
}
\arguments{
\item{dt.ad}{- An S3 object of MVAD type (partially evaluated)}
}
\value{
An S3 object of MVAD type (partially evaluated).
}
\description{
\code{preprocess_data} returns the input data normalized and detrended (unitless).
Before calling, \code{\link{setup_model}}, and\code{\link{add_data}} should be called.
}
