% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{setup_model}
\alias{setup_model}
\title{Set initial parameters for MVAD.}
\usage{
setup_model(
  num_variables = 4,
  tau = 10 * 60 * 30,
  data_resolution = 15,
  num_lag_terms = 1
)
}
\arguments{
\item{num_variables}{Integer - defining ncol of training/testing data tables.}

\item{tau}{Integer - refering to number of data points to train model over.
i.e., 10 mins @ 30Hz  = 10 * 60 * 30.}

\item{data_resolution}{Integer - number of terms to be 'smoothed' for per-point by the model estimation.}

\item{num_lag_terms}{Integer - number of lag terms to include in the model. TODO expand from 1 to k}
}
\value{
An S3 object of MVAD type.
}
\description{
\code{setup_model} returns the initial mvad object created be initializing parameters.
}
\examples{
dt.ad <- setup_model(4, 10 * 60 * 30, 15, 1)
dt.ad <- setup_model() ## defaults to above

}
