% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{train_model}
\alias{train_model}
\title{Train Model.}
\usage{
train_model(dt.ad)
}
\arguments{
\item{dt.ad}{- An S3 object of MVAD type (partially evaluated)}
}
\value{
An S3 object of MVAD type (partially evaluated).
}
\description{
\code{train_model} estimates a VAR(k) model fitted on the preprocessed training data.
Before calling, \code{\link{setup_model}}, \code{\link{add_data}}, and \code{\link{preprocess_data}} should be called.
}
